Created a basic demo involving ball in a box
Got tired brain to work.
Came up with something that works.
Wow it's a haiku.

The red ball is bouncing in a white box. The foundations I laid earlier means now the demo is actually running a physics engine. I can also shake the screen. Problem is that this scene took me 200 lines of code to configure. Going to need to extract that into classes somehow.